dnsmasq: fix uci-defaults script to exit 0 so it is cleaned up
A file, package/network/services/dnsmasq/files/50-dnsmasq-migrate-resolv-conf-auto.sh,
was added in commit 6a2855212096d2c486961a0841b037bae4b75de7, but it
does not exit in a way that tells the uci-defaults mechanism that it
succeeded, and so it is not cleaned up after running successfully. Add
an exit 0 to the end to correct that.

Signed-off-by: Russell Senior <russell@personaltelco.net>
(cherry picked from commit 731f7ea48a8a1f62de1a63e52b2ed8e245f12ac3)

Change-Id: I8ffd35e3dae9c2a25be54888c984fb4fb6e5b89a
Signed-off-by: Madhurya Kandamuru <quic_madhurya@quicinc.com>