netifd: check whether module load is done
We observed that network modules are getting loaded after
the network service is run, this is creating some race
between network modules and bridge interface creation.

As a workaround we added a logic to wait for module load
to complete so that network service creates the bridge
interfaces after all network modules are loaded.

Change-Id: I25acdbd149f36bdb3d31d22e5cd58e86f410b96a
Signed-off-by: Ram Chandra Jangir <quic_rjangir@quicinc.com>