verbose.mk: add V=e option to show start and stop for tasks
Sometimes when running multiple jobs in parallel one of them will
fail and it's difficult to determine which one. This adds a new
V=e option to track when a job starts and stops. You can invoke
via:

 $ make -j24 V=e

You can run something like this now to see when each task begins
and completes:
 $ make -j24 V=e toolchain/uClibc/compile | awk '{ print strftime("%H:%M:%S"), $0; }'

without this change, for target/linux/prereq, SUCCESS_MESSAGE
definition is not available leading to compilation issue
when we make V=e.

Change-Id: I7303cdb17c7a1ee710be4e05d108d12219dd88f1
Signed-off-by: Atchaya R <quic_ratchaya@quicinc.com>