fossid scan: FOSSID scan support in QSDK builds
This change enables FOSSID scan to be triggered in parallel
with the QSDK build. This can be enabled by setting the config
'CONFIG_FOSSID_SCAN=s' during the build along with make -
'make V=ew -j16 CONFIG_FOSSID_SCAN=s'

Setting this config runs the scan in parallel to the build, sends the
generated output JSON from the FOSSID to the VSF API for the report
generation which converts the JSON into Excel report.

If the developer is compiling a single package and wants the
FOSSID scan report for that package, make command will be -
'make <package_path> V=ew CONFIG_FOSSID_SCAN=sr'

The credentials to access the VSF API through curl will be
read from the netrc file.

Change-Id: I2609f60a4ab7461eca2421f9d559ab3a3ea91cca
Signed-off-by: Madhurya Kandamuru <quic_madhurya@quicinc.com>