hostapd.sh: Add EHT80 to htmode to avoid disabling vht
During Mesh configuration, disable_vht=1 option is added
if HE/VHT/EHT 20/40/160 and HE/VHT80 is NOT set. So, if
user configures in EHT80 mode, because of the upgraded
script disable_vht option is set, hence Mesh will always
brought up in EHT40 instead of EHT80.

Fix this by adding EHT80 case in supported ht_mode, but
keep the case to do nothing as no special handling is
needed now.

Change-Id: I965fc5601557f31d7e3d87cec43588e217f9231a
Signed-off-by: Manish Dharanenthiran <quic_mdharane@quicinc.com>