mac80211.sh: Set default channel as 6 for 2G instead of 1
Set the default channel as 6 for 2G instead of 1. If channel 1
is set, the script is getting the channel list for 2G as 1-233
instead of 1-14 with split phy is enabled.

This is because, the probing sequence of radios are changed and
the loop to get the channel range for 2G breaks with 6G channel
list as it contains channel 1.

Hence, fix this issue by setting the 2G radio's default channel
as 6.

Change-Id: Ida729b0f7c7f354602b5c47fb92beb9587c18c3b
Signed-off-by: Maharaja Kennadyrajan <quic_mkenna@quicinc.com>