mac80211.sh: Fix interface cleanup
Currently we are not killing pid process in case if devices name would be 6
characters for non-single wiphy architecture, due to which hostapd is still
running in background. Also removal of lock file is not handled.

Fix this issue by passing proper argument to the function and deleting lock
file.

Instead of kill -9, we should use kill -15 to kill a process gracefully.

Add freq_list for mesh mode.

Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>
Change-Id: Iea4baa5506fd299bb415e2eea1289336a2fb047e