mac80211.sh: destroy MLO hostapd instance upon reload
Currently whenever network reload is given, the previously created single
instance of MLO hostapd is not destroyed. This leads to issue in bringing
up the new instance since it will use the same interfaces. Sometimes,
interfaces will not be even up.

Fix this issue by properly destroying the previous instance when reloading.

Change-Id: I460dd021e42754923ef0a01c0e7a418ff3d78743
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>