mac80211.sh: Fix interface not coming up after network restart
Currently, interfaces are not comimg up after network restart is given.

This is because we are stuck in infinite while loop in func
mac80211_setup_supplicant for mesh interface.

Also after running hostapd command to bring interfaces up, we should wait for
5g interface untill DFS is done then run wireless_add_process.

Change sequence of clean up because hostapd teardown is taking time and due
to this wpa_supplicant clean up is not properly done. Move hostapd teardown
in last.

Add missing function mac80211_set_fq_limit in script.
Fixes: I228b02b9f3b3 ("wifi: script: Port DS related changes")

Change-Id: I09a64bdb6f16bb67c3a92386c53d929ddbbeaea6
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>
Signed-off-by: Maharaja Kennadyrajan <quic_mkenna@quicinc.com>