wifi: mac80211.sh: Adding condition to stop infinite loop
There is workaround where loop is continued after the invalid case but
failed to increment the bandidx in the script which caused the AP to
hold procd for so long. This results in failure in the
firmware loading during insmod of ath12k module. This invokes firmware
timeout of 60 seconds which eventually fails the coldboot and trigger
the kernel stack corruption.

Change-Id: I068813b1d4a62157e516f846c0851ccd5c47fb5e
Signed-off-by: Gautham Kumar Senthilkumaran <quic_gauthamk@quicinc.com>
(cherry picked from commit 60f9cbb637bccd53211aaa1afdda861dcc5465ec)