mac80211.sh: Fix multiple bssid for HE htmode
Currently when we try to enable multiple mbssid in 2G and 5G with He mode,
interfaces are not coming up.

This is because we are not adding multiple_bssid for HE, both HE and EHT allows
multiple bssid, but in EHT we don't support it on lower bands (2.4/5GHz) hence
the script restricts it.

Change-Id: I413659c8209cbde5e55f6a77cd164e17ee5e4ff9
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>