mac80211.sh: Fix for dhcp issue with mesh config
mac80211.sh script file is stuck in a loop while trying to get
the link id.

There are two reason for this stuck.
1. Hard coded interface value
2. Trying to get link id before interface creation

Due to this hard coded value, it is failed to get link id and
loop was added until to get the valid link id.

Sometimes, the interface creation is getting delayed and failed
to get the link id due to this invalid interface.

Removed the hard code interface name and added proper iface
name to get the link id.

Avoided to get the link id if there is no valid interface.

Fixes: e8f58d4d5dff2(Fix interface not coming up after network
restart)

Change-Id: I2ee73f5a45c6cd710a3eea36872a61fab06ca16a
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>