mac80211.sh Add the primary_link selection option
Add the support to configure the primary_link on /etc/
config/wireless which helps to select the primary link manually
file and the script has dynamically update the value in the
debug_fs based on the link-id in the wifi config file.

Usage:

Configure the primary_link in the wifi config file as:

config wifi-mld '<mld_name>'
	option primary_link '<link_id>'

Possible link_ids are 0,1 and 2.

Note: While configure the primary_link, it is must
      to set the ifname in the wifi-mld params.

Change-Id: I1661e883924cc308fac94ba814578752db0fc13b
Signed-off-by: Arulanbu Balusamy <quic_abalusam@quicinc.com>