mac80211.sh: use mac-address per hardware as reference in single-wiphy
Currently, in sigle-wiphy case only one mac-address is read from
ieee80211 sysfs. This mac-address is used as a reference to derive
mac addresses for all the wifi interface. Thus, the wifi interface
created on different hardware/SoC get the mac-address assigned
which is based on single mac-address. That means not all the wifi
interface uses the correct mac-address of the underlying hardware
in single-wiphy.

Thus, add support to read mac-addresses for all the underlying
hardware in single-wiphy. Also, use the correct mac-address of the
underlying hardware as a reference to derive the mac-address for an
interface in the same hardware.

Change-Id: Ieb96cac5c706d8805616154345b99a3194001c69
Signed-off-by: Raj Kumar Bhagat <quic_rajkbhag@quicinc.com>