mac80211.sh: Rectify eht_oper_centr_freq_seg0_idx for EHT40
When an AP is brought up on HT40PLUS or HT40MINUS in 6 GHz, the
output of " hostapd_cli -i wlanx status" shows
eht_oper_centr_freq_seg0_idx=0.

Fix this issue, by assigning eht_center_seg0 to $idx, in case of EHT40
within mac80211_hostapd_setup_base, if hwmode is a.

Also, if hwmode is a and if htmode is HT20, assign eht_center_seg0 and
vht_center_seg0 to $idx.

Change-Id: I16fbfef0a12731978d731a81fbfc8e691a34112e
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>