mac80211.sh: Fix interface not coming up after network restart
Currently, whenever network reload is given, the previously created
single instance of MLO hostapd is not destroyed and interfaces are
not coming-up. When 2G interface is deleted MLO hostapd is not destroyed
because the file name of MLO hostapd instance is suffixed with
device name.

Fix this issue by changing the hostapd pid file name with phy.

Change-Id: I71fe63535ed1161e8c11b2bb77e92e95b54f0d81
Signed-off-by: Harish Rachakonda <quic_rachakon@quicinc.com>