openwrt : blacklist legacy iptables
The following warning print is observed with iptables command.
"#Warning: iptables-legacy tables present, use iptables-legacy to see them"

The iptables binary throws this warning if legacy iptable modules
are present in the system.

Openwrt 23.05 and greater use nf tables instead of legacy iptables.

Since the legacy iptables modules are not used in the system,
disable the autoload of those modules.

Change-Id: I607cc5429d58579819b4a9111101078e626ae0a5
Signed-off-by: Hari Chandrakanthan <quic_haric@quicinc.com>