ar71xx: fix saving config after sysupgrade in nand
The changes for upgrade linux kernel to v5.4.
On sysupgrade in nand, the /etc/config changes are not preserved.
After sysupgrade completes and reboots, it is considered as first_boot
and overlay is deleted. So, instead of untar the config tarball before
reboot, copy it into the overlay. While mounting root during the preinit,
the untar of the tarball will be done and the configs are restored.

Change-Id: I718e84795849c4400034bb16b222f326d92a7a2c
Signed-off-by: Ronghua Zeng <quic_ronghuaz@quicinc.com>