Makefile: Added new Makefile path for downgrade of WLAN components
WLAN components needed to be downgraded for SOD branch. Hostapd,backport and
iw version are downgraded as similar to the mainline branch. This downgrade
required new Makefile to point to the older version of those components.

These new Makefile is already present in the each WLAN components.
But after moved to new Openwrt version, main Makefile needed modification
to fetch that new Makefile for downgrade. This is already present in the
previous branch "Hazelnut_dev", now ported the same changes to new
branch "Hazelnut". Added Marina related code changes in the Makefile for
compilation.

Change-Id: I48f7c14f2f01ea5dfa328f3422ea1bcf67a963d7
Signed-off-by: Gautham Kumar Senthilkumaran <quic_gauthamk@quicinc.com>