mac80211.sh: Add support for 6GHz low and 6GHz high
The current script doesn't generate the config for 6GHz Low
and 6GHz High radios are present in a board and it will generate
the wireless config with 6GHz full band frequencies for 1 of the 6GHz
radio and won't generate the config for another 6GHz radio.

Hence, added the support to generate the wireless config for 6GHz low
and high with proper frequency ranges.

Change-Id: Ife2ce918c32ccd545f00a757dea4815c81037729
Co-authored-by: Rajat Soni <quic_rajson@quicinc.com>
Signed-off-by: Maharaja Kennadyrajan <quic_mkenna@quicinc.com>